Fix socket options on Linux
